Move old expr_math_op into nask_utility namespace
