+++ src/string_util.cpp
@@ -1,6 +1,5 @@
 #include "string_util.hpp"
 #include "mod_rm.hpp"
-#include "tinyexpr.h"
 #include "nask_defs.hpp"
 
 namespace nask_utility {
@@ -65,27 +64,6 @@ namespace nask_utility {
             }
             return tmp;
 
-        } catch (std::regex_error& e) {
-            log()->info(e.what());
-        }
-        return "";
-    }
-
-    std::string expr_math_op(const std::string& subject) {
-
-        const std::regex reg_numeric("([^0-9]*)([-\\*/+0-9]*)([^0-9]*)");
-
-        try {
-            std::smatch match;
-            if (std::regex_search(subject, match, reg_numeric) && match.size() > 1) {
-                int error;
-                const int process = te_interp(match[2].str().c_str(), &error);
-                const std::string cat = match[1].str() + match[2].str() + match[3].str();
-                const std::string empty = "";
-
-                std::string result = replace(subject, cat, empty);
-                return match[1].str() + std::to_string(process) + match[3].str() + result;
-            }
         } catch (std::regex_error& e) {
             log()->info(e.what());
         }
